refactor: break CustomSelect circular deps
